ref(py3): Remove set_exception_info from subclassed Future
Our requirements file loads the backported pythonfutures library ONLY in
python2. In python3 this is part of the standard library.

Unfortunately the python3 implementation does NOT have a
`set_exception_info` method on the `Future` object, only `set_exception`
which does not keep track of the traceback.

In fact, this was never part of PEP-3148 [0], however was added into the
backport library in [1], which we implemented in our subclassed futures.

[0]: https://www.python.org/dev/peps/pep-3148/
[1]: https://github.com/agronholm/pythonfutures/commit/6342a77a56fa04ac9c9b28cbf4d1ac26be3f0476

The implication of this is that futures which raise exceptions will now
unfortunately lose the traceback.